+++ packages/isaac-typescript-definitions/src/types/classes/Color.d.ts
@@ -57,10 +57,5 @@ declare interface Color extends IsaacAPIClass {
 declare namespace Color {
   function Lerp(m1: Color, m2: Color, t: float): Color;
 
-  /**
-   * @deprecated This can be mutated by other mods, so it is never safe to use. Use the
-   *             `ColorDefault` constant from the standard library instead. Alternatively, you can
-   *             create your own constant that is local to your own mod.
-   */
-  const Default: never;
+  // The `Default` preset is not implemented, since it is unsafe. See the `ColorDefault` constant.
 }